Refatorado tela de edição de usuários

+++ resources/js/views/UsersEdit.vue
@@ -1,34 +1,33 @@
 <template>
     <div>
-        <div v-if="message" class="alert">{{ message }}</div>
         <div v-if="!loaded">Loading...</div>
-        <form @submit.prevent="onSubmit($event)">
+        <form @submit.prevent="onSubmit($event)" class="col-6 mt-3">
             <div class="form-group">
-                <label for="user_name">Name</label>
-                <input id="user_name" v-model="user.name"/>
+                <label for="user_name">Nome</label>
+                <input id="user_name" v-model="user.name" class="form-control" placeholder="Seu nome" required/>
             </div>
             <div class="form-group">
-                <label for="user_email">Email</label>
-                <input id="user_email" type="email" v-model="user.email"/>
+                <label for="user_email">E-mail</label>
+                <input id="user_email" type="email" v-model="user.email" class="form-control" placeholder="Seu e-mail" required/>
             </div>
             <div class="form-group">
                 <label for="user_telefone">Telefone</label>
-                <input id="user_telefone" type="phone" v-model="user.telefone"/>
+                <the-mask :mask="['(##) #####-####']" id="user_telefone" type="phone" v-model="user.telefone" class="form-control" required/>
             </div>
             <div class="form-group">
-                <button type="submit" :disabled="saving">Update</button>
-                <button :disabled="saving" @click.prevent="onDelete($event)">Delete</button>
+                <button type="submit" :disabled="saving" class="btn btn-secondary">Update</button>
+                <button :disabled="saving" @click.prevent="onDelete($event)" type="button" class="btn btn-danger">Delete</button>
             </div>
         </form>
     </div>
 </template>
 <script>
     import api from '../api/users';
+    import Swal from 'sweetalert2';
 
     export default {
         data() {
             return {
-                message: null,
                 loaded: false,
                 saving: false,
                 user: {
@@ -43,26 +42,66 @@
             onSubmit(event) {
                 this.saving = true;
 
+                if (this.user.telefone.length < 11) {
+                    Swal.fire({
+                        type: 'error',
+                        title: 'Oops...',
+                        text: 'Preencha o campo de telefone corretamente!'
+                    });
+
+                    this.saving = false;
+                    return;
+                }
+
                 api.update(this.user.id, {
                     name: this.user.name,
                     email: this.user.email,
                     telefone: this.user.telefone,
                 }).then((response) => {
-                    this.message = 'User updated';
-                    setTimeout(() => this.message = null, 2000);
-                    this.user = response.data.data;
+                    Swal.fire({
+                        position: 'top',
+                        type: 'success',
+                        title: 'Usuário salvo!',
+                        showConfirmButton: false,
+                        timer: 1500
+                    });
+                    this.$router.push({ name: 'users.index' });
                 }).catch(error => {
-                    console.log(error)
+                    Swal.fire({
+                        type: 'error',
+                        title: 'Oops...',
+                        text: 'Erro ao salvar o usuário!',
+                        footer: error.message
+                    });
                 }).then(_ => this.saving = false);
             },
             onDelete() {
-              this.saving = true;
+                this.saving = true;
 
-              api.delete(this.user.id)
-                 .then((response) => {
-                     this.message = 'User Deleted';
-                     setTimeout(() => this.$router.push({ name: 'users.index' }), 2000);
-                 });
+                Swal.fire({
+                    title: 'Você tem certeza?',
+                    text: "Você não poderá reverter isso!",
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Sim, exclua-o!',
+                    cancelButtonText: "Cancelar"
+                }).then((result) => {
+                    if (result.value) {
+                        api.delete(this.user.id)
+                            .then((response) => {
+                                Swal.fire(
+                                    'Deletado!',
+                                    'Usuário foi deletado.',
+                                    'success'
+                                );
+                                this.$router.push({ name: 'users.index' });
+                            });
+                    }
+                }).finally(_ => {
+                    this.saving = false;
+                });
             }
         },
         created() {
@@ -76,19 +115,3 @@
         }
     };
 </script>
-<style lang="scss" scoped>
-    $red: lighten(red, 30%);
-    $darkRed: darken($red, 50%);
-    .form-group label {
-        display: block;
-    }
-    .alert {
-        background: $red;
-        color: $darkRed;
-        padding: 1rem;
-        margin-bottom: 1rem;
-        width: 50%;
-        border: 1px solid $darkRed;
-        border-radius: 5px;
-    }
-</style>
